# contracts/openapi/openapi_v1.0.6_enveloped.yaml
openapi: 3.0.3
info:
  title: Hourglass Rentals API
  version: 1.0.6
  description: '面向西班牙（Spain）上线的租赁管理平台 API（Admin / Owner / Tenant + Public Website + CRM）。

    本文件是字段级接口“唯一真相源（Source of Truth）”。前后端均以此为准。


    约定：

    - 所有需要翻译展示的文本字段同时保留原文（*_original）并提供翻译缓存（*_i18n）与展示字段（*_display）。

    - 服务端根据用户偏好与请求头 Accept-Language 解析 viewerLanguage，并填充 *_display。

    - 所有时间使用 ISO 8601，货币默认 EUR。



    ---


    **M0 冻结说明（2025-12-24）**

    - 生产统一使用 `/v1`（server url 已带 `/v1`）；原 `/api` 仅作为短期兼容（可由网关 rewrite）。

    - 统一响应 Envelope：成功响应为 `{requestId,timestamp,data}`；错误响应为 `{requestId,timestamp,error}`。



    ### M0-FULL 补丁说明（v1.0.2）

    - 补齐：/health、/invitations、/gdpr、property images upload

    - 安全：补齐通用错误码（401/403/404/429/500）与 Cookie+CSRF 口径

    - 兼容：/api → /v1 下线计划见 `00_RELEASE/API_兼容与下线计划.md`



    ### M0-SUPER 补丁说明（v1.0.3）


    - 新增：`GET /v1/gdpr/jobs/{jobId}`（GDPR job 查询，便于前端轮询）

    - 新增：`POST /v1/auth/refresh`、`POST /v1/auth/logout`（Cookie 会话刷新/退出）

    - 安全：补充 refresh cookie 的可执行表达（`refreshCookie` security scheme）



    v1.0.4（M0-Ultimate）变更摘要：

    - GDPR：新增历史任务列表/筛选接口 `GET /gdpr/jobs`（支持 type/status/target 过滤 + cursor 分页）。

    - 并发：引入 If-Match / ETag 乐观锁约定（rowVersion 全链路），写接口可选强制门禁（412/428）。


    补充：ErrorResponse.error.code 已固定为 ErrorCode 枚举（前后端统一）。


    【v1.0.6 重要变更（M0修正）】

    - 金额字段：对外统一使用字符串（MoneyString），避免 JSON number 精度问题；落库使用 NUMERIC(18,2)。

    - Document 新增 scanStatus 字段（UPLOADED/SCANNING/READY/BLOCKED/FAILED）。

...

    '
  contact:
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '建议使用 Google Identity Platform / Firebase Auth / Auth0 等签发 JWT；API 侧只做校验与 RBAC。


        ⚠️ M0-FULL：浏览器端推荐使用 HttpOnly Cookie 会话（见 sessionCookie + CSRF），Bearer 仅建议用于服务到服务或本地调试。'
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly Secure Session Cookie（浏览器端默认认证方式）。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF 防护 token（对所有写接口 POST/PUT/PATCH/DELETE 必须）。
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh
      description: HttpOnly Secure Refresh Cookie（用于会话刷新/轮换）。
  parameters:
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        example: zh-CN,es;q=0.9,en;q=0.8
      description: 期望返回的展示语言。服务端可结合用户偏好与此头决定 viewerLanguage。
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: 乐观锁：客户端从响应头 ETag 获取并在更新/状态变更/删除时通过 If-Match 回传。推荐格式：W/"{rowVersion}".
  schemas:
    ErrorCode:
      type: string
      description: 统一错误码枚举（用于 ErrorResponse.error.code）。前后端/多端必须保持一致。
      enum:
      - VALIDATION_ERROR
      - BAD_REQUEST
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - CONFLICT
      - PRECONDITION_REQUIRED
      - PRECONDITION_FAILED
      - RATE_LIMITED
      - INTERNAL_ERROR
    ErrorResponse:
      type: object
      required:
      - error
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          description: 请求追踪 ID（用于排障）
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            details:
              type: object
              additionalProperties: true
        timestamp:
          type: string
          format: date-time
          description: 服务端生成的响应时间戳
...
    I18nText:
      type: object
      description: 机器翻译缓存结构（用于用户输入文本：工单、通知、资料请求备注等）。
      required:
      - originalText
      - originalLang
      properties:
        originalText:
          type: string
          maxLength: 20000
        originalLang:
          $ref: '#/components/schemas/LanguageCode'
        translations:
          type: object
          description: key=LanguageCode，value=翻译文本
          additionalProperties:
            type: string
            maxLength: 20000
        displayText:
          type: string
          readOnly: true
          description: 服务端根据 viewerLanguage 解析后的展示文本。
        provider:
          type: string
          description: 翻译供应商/引擎标识（可选）
        model:
          type: string
          description: 模型/版本（可选）
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        updatedAt:
          type: string
          format: date-time
...
    TicketStatus:
      type: string
      enum:
      - PENDING_ADMIN
      - REJECTED_ADMIN
      - WAITING_LANDLORD
      - IN_PROGRESS
      - DONE
      - CANCELLED
      - NEW
      description: NEW 为 admin 原型兼容值；服务端应标准化为 PENDING_ADMIN。
    TicketPriority:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - NORMAL
...
    Ticket:
      type: object
      required:
      - id
      - code
      - propertyId
      - priority
      - status
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
        code:
          type: string
        ownerId:
          type: string
          nullable: true
        tenantId:
          type: string
          nullable: true
        propertyId:
          type: string
        leaseId:
          type: string
          nullable: true
        title:
          type: string
        title_i18n:
          allOf:
          - $ref: '#/components/schemas/I18nText'
          nullable: true
        title_display:
          type: string
          readOnly: true
          nullable: true
        description:
          type: string
          nullable: true
        description_i18n:
          allOf:
          - $ref: '#/components/schemas/I18nText'
          nullable: true
        description_display:
          type: string
          readOnly: true
          nullable: true
        category:
          type: string
          nullable: true
          description: 可选枚举：plumbing/electrical/appliance/internet/other（保持开放以免阻塞）
        priority:
          $ref: '#/components/schemas/TicketPriority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        estCost:
          allOf:
          - $ref: '#/components/schemas/MoneyString'
          nullable: true
        estCostRange:
          allOf:
          - $ref: '#/components/schemas/TicketCostRange'
          nullable: true
        assignee:
          type: string
          nullable: true
        vendor:
          type: string
          nullable: true
        quoteAmount:
          allOf:
          - $ref: '#/components/schemas/MoneyString'
          nullable: true
        quoteAttachmentIds:
          type: array
          items:
            type: string
          nullable: true
        approvalStatus:
          allOf:
          - $ref: '#/components/schemas/TicketApprovalStatus'
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        actualCost:
          allOf:
          - $ref: '#/components/schemas/MoneyString'
          nullable: true
        linkedExpenseId:
          type: string
          nullable: true
        resultNote:
          type: string
          nullable: true
        resultNote_i18n:
          allOf:
          - $ref: '#/components/schemas/I18nText'
          nullable: true
        resultNote_display:
          type: string
          readOnly: true
          nullable: true
        attachmentIds:
          type: array
          items:
            type: string
          default: []
        comments:
          type: array
          items:
            $ref: '#/components/schemas/TicketComment'
          default: []
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        rowVersion:
          type: integer
          format: int64
          description: 用于乐观锁的行版本号（与 ETag/If-Match 对应）。
...
    PresignUploadRequest:
      type: object
      required:
      - fileName
      - mime
      - size
      - docType
      - visibility
      properties:
        fileName:
          type: string
        mime:
          type: string
        size:
          type: integer
          minimum: 1
        sha256:
          type: string
          nullable: true
        docType:
          $ref: '#/components/schemas/DocumentDocType'
        expiresAt:
          type: string
          format: date
          nullable: true
        visibility:
          $ref: '#/components/schemas/DocumentVisibility'
        links:
          type: object
          description: 绑定关系（至少提供一个）
          properties:
            ownerId:
              type: string
              nullable: true
            tenantId:
              type: string
              nullable: true
            propertyId:
              type: string
              nullable: true
            leaseId:
              type: string
              nullable: true
            invoiceId:
              type: string
              nullable: true
            ticketId:
              type: string
              nullable: true
            entityType:
              type: string
              nullable: true
            entityId:
              type: string
              nullable: true
          additionalProperties: false
...
    CompleteUploadRequest:
      type: object
      required:
      - uploadId
      - docId
      properties:
        uploadId:
          type: string
        docId:
          type: string
        size:
          type: integer
          nullable: true
        sha256:
          type: string
          nullable: true
    EnvelopeMeta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          description: 请求追踪 ID（用于排障）
        timestamp:
          type: string
          format: date-time
          description: 服务端生成的响应时间戳
...
    Envelope_Ticket:
      allOf:
      - $ref: '#/components/schemas/EnvelopeMeta'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/Ticket'
...
    Envelope_PresignUploadResponse:
      allOf:
      - $ref: '#/components/schemas/EnvelopeMeta'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/PresignUploadResponse'
    Envelope_Document:
      allOf:
      - $ref: '#/components/schemas/EnvelopeMeta'
      - type: object
        required:
        - data
        properties:
          data:
            $ref: '#/components/schemas/Document'
...
  responses:
    BadRequest400:
      description: 请求参数错误 / 校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 未登录或会话失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden403:
      description: 无权限（或策略要求返回 404 以防止资源枚举）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound404:
      description: 资源不存在（或越权时按 404 处理）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict409:
      description: 并发冲突 / 乐观锁失败（expected_row_version/If-Match）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests429:
      description: 请求过于频繁（限流触发）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError500:
      description: 服务端异常
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionFailed412:
      description: Precondition Failed（ETag/If-Match 不匹配，资源已被他人修改）
      content:
        application/json:
          schema: &id001
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionRequired428:
      description: Precondition Required（缺少 If-Match；该接口启用强制乐观锁门禁）
      content:
        application/json:
          schema: *id001
  headers:
    ETag:
      description: 资源版本 ETag（由 rowVersion 派生）。用于 If-Match 乐观锁。
      schema:
        type: string
paths:
...
  /tickets/{ticketId}:
    get:
      tags:
      - Tickets
      summary: 获取工单详情
      security:
      - sessionCookie: []
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Ticket'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
      x-rate-limit-tier: STANDARD
      x-error-strategy: 403/404
      operationId: getTicketsByTicketId
      description: 获取工单详情
    patch:
      tags:
      - Tickets
      summary: 更新工单（Admin/Owner，或 Tenant 修改自己工单的可编辑字段）
      description: '当 status=DONE 时，服务端必须校验 resultNote 非空，且具备至少 1 个附件或 linkedExpenseId（与原型一致）。


        并发控制：该接口启用 If-Match 乐观锁（基于 rowVersion/ETag）。'
      security:
      - sessionCookie: []
        csrfToken: []
      parameters:
      - name: ticketId
        in: path
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/AcceptLanguage'
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ticket'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Ticket'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '409':
          $ref: '#/components/responses/Conflict409'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
        '412':
          $ref: '#/components/responses/PreconditionFailed412'
        '428':
          $ref: '#/components/responses/PreconditionRequired428'
      x-rate-limit-tier: STANDARD
      x-audit: true
      x-audit-action: TICKET_UPDATE
      x-error-strategy: 403/404
      operationId: patchTicketsByTicketId
...
  /files/presign-upload:
    post:
      tags:
      - Files
      summary: 获取直传对象存储的签名上传 URL（推荐）
      security:
      - sessionCookie: []
        csrfToken: []
      parameters:
      - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PresignUploadRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_PresignUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
      x-rate-limit-tier: STANDARD
      x-error-strategy: 403/404
      operationId: postFilesPresignUpload
      description: 获取直传对象存储的签名上传 URL（推荐）
  /files/complete-upload:
    post:
      tags:
      - Files
      summary: 完成上传并落库 Document 元数据
      security:
      - sessionCookie: []
        csrfToken: []
      parameters:
      - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteUploadRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_Document'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '403':
          $ref: '#/components/responses/Forbidden403'
        '404':
          $ref: '#/components/responses/NotFound404'
        '429':
          $ref: '#/components/responses/TooManyRequests429'
        '500':
          $ref: '#/components/responses/InternalError500'
      x-rate-limit-tier: STANDARD
      x-error-strategy: 403/404
      operationId: postFilesCompleteUpload
      description: 完成上传并落库 Document 元数据
...